+++ docs/history.md
@@ -3,6 +3,7 @@
   * With command line and front end
 * Adding enable / disable sprint config
 * Adding enable / disable dig (!caution the bot can stucks)
+* Adding reloading configuration button
 
 
 ## 0.1.6